+++ src/main/java/pl/asie/computronics/oc/driver/RobotUpgradeSpeech.java
@@ -157,6 +157,7 @@ public class RobotUpgradeSpeech extends ManagedEnvironmentWithComponentConnector
 		super.onDisconnect(node);
 		if(node == this.node()) {
 			isValid = false;
+			stopTalking();
 		}
 	}